facedetection/requirements.txt
grpcio==1.35.0
grpcio-tools==1.35.0
opencv-python==4.5.2.54
imageio==2.9.0
imageio-ffmpeg==0.4.4
numpy==1.20.3
Cython==0.29.23
mxnet==1.8.0.post0 # change for cuda version if gpu should be used
torch==1.9.0 # change for cuda version if gpu should be used
scipy==1.7.0
PyYAML==5.3.1
